Refactorig
* Add generic validation.
* Move ToOptional to a new utils.h header.
